requirements.txt
# LangSmith for tracing and observability
langsmith>=0.1.0

# Evaluation frameworks
ragas>=0.1.0
deepeval>=0.21.0

# LLM providers
openai>=1.0.0
anthropic>=0.18.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
plotly>=5.18.0

# Configuration and validation
pydantic>=2.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
